[스프링 - 협업 미션] 로키(김경록) 미션 제출합니다. (#44)
* chore: 서버 환경설정 세팅

* chore: 테이블 세팅 변경

* chore: 어플리케이션 세팅 변경

* chore: db 초기화 기능 추가

* chore: DB 제거 순서 변경

* feat: 중복 가입 요청시 bad request 응답

* refactor: controller final 처리 및 제네릭 추가

* feat: 전체 조회 기능 추가

* feat: 요금 조회 기능 추가

* feat: 정보 수정시 로그인 검증 기능 추가

* fix: Line 제거 오류 수정

- StationDao exists 쿼리 추가

* feat: 노선 색상 중복 방지 기능 추가

* fix: 역 id 정상 증가

* fix: preflight 설정 추가 (CORS)

* feat: 존재하지 않는 역 관련 예외 추가

* feat: 유효하지 않은 노선 요청 검증 기능 추가

* feat: 유효하지 않은 역과 구간 요청 검증 기능 추가

* feat: 노선에 등록된 역 삭제 방지 기능 추가

* fix: 구간 관련 에러 처리

* refactor: 테스트 관련 처리

* feat: NotFoundException 추가

* refactor: 테스트시 로그인 토큰 전달

* refactor: id 순으로 역 정보 출력

* test(Station): 토큰이 없거나 유효하지 않은 경우에 대한 테스트 추가

* fix: 예외 발생시 반환 상태코드 변경

* chore: swagger 의존성 추가

* feat: api 문서 추가 (Swagger)

* feat: 멤버 나이별 할인 금액 구현

* feat: 노선 추가 요금 기능 구현

- 나이별 할인 방식 변경
- 패키지 정리

* refactor: Price 클래스 Fare로 리팩토링

* refactor: 요금 계산 방식 리팩토링

* fix: 테스트 수정

* test: 서비스 레이어 테스트 추가

- 애플리케이션 시작시 배너 변경

* refactor: 피드백 반영

* refactor: 요금 계산 로직 Fare의 역할로 변경

* refactor: 중복되는 예외 메시지, 디폴트 메시지로 변경

* refactor: 사용하지 않는 코드 제거

* refactor: 주석 제거

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,29 +1,36 @@
 package wooteco.subway.station.application;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service
 public class StationService {
-    private StationDao stationDao;
 
-    public StationService(StationDao stationDao) {
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
+
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
+        if (stationDao.exists(stationRequest.getName())) {
+            throw new DuplicatedStationException(stationRequest.getName());
+        }
         Station station = stationDao.insert(stationRequest.toStation());
         return StationResponse.of(station);
     }
 
     public Station findStationById(Long id) {
-        return stationDao.findById(id);
+        return stationDao.findById(id)
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 StationID 입니다. id: " + id));
     }
 
     public List<StationResponse> findAllStationResponses() {
@@ -35,6 +42,9 @@ public class StationService {
     }
 
     public void deleteStationById(Long id) {
+        if (sectionDao.existsByStationId(id)) {
+            throw new InvalidDeletionException("이미 노선에 등록되어있는 역은 지울 수 없습니다.");
+        }
         stationDao.deleteById(id);
     }
 }
